The class Length has been created.
The class Length is a subclass of Unit, where it has the methods to change from metres to yards and to milles. On Unit, it has the method to_unit for verifying if the unit received exists. Unless it exists, it returns an error. On Length, it has the method __str__ which adds 'metres', 'm' or nothing for gaiving the number and its unit.